chore: removed extensions folder and switch all imports (+ small fixes)
