+++ src/commands/accents-setter/consts.ts
