+++ src/commands/accents-setter/quick-pick.ts
