+++ src/commands/index.ts
@@ -1,4 +1,2 @@
 export {default as accentsSetter} from './accents-setter';
 export {default as accentsQuickPick} from './accents-setter/quick-pick';
-export {default as fixIcons} from './theme-icons';
-export {default as toggleApplyIcons} from './toggle-apply-icons';
