+++ src/consts/paths.ts
@@ -3,12 +3,11 @@ import * as path from 'path';
 import {IPaths} from '../interfaces/ipaths';
 
 export const PATHS: IPaths = {
-  DIST: './dist',
-  ICONS: './icons',
   SRC: './src',
-  THEMES: './themes',
-  UI: './ui',
-  VSIX_DIR: path.join(__dirname, '../../'),
+  THEMES: './out/themes',
+  UI: './out/ui',
+  VSIX_SRC_DIR: path.join(__dirname, '../..'), // From "src" dir
+  VSIX_DIR: path.join(__dirname, '../../..'), // From "out" dir
   EXT_DIR: path.join(__dirname, '..')
 };
