+++ src/helpers/check-installation.ts
