+++ src/helpers/write-changelog.ts
@@ -1,10 +1,10 @@
 import * as path from 'path';
 import {getDefaultValues, getPackageJSON, writeFile} from './fs';
 
-import {IDefaults} from './../interfaces/idefaults';
+import {IDefaults} from '../interfaces/idefaults';
 
 const writeDefaults = (defaults: IDefaults) =>
-  writeFile(path.join('./extensions/defaults.json'), JSON.stringify(defaults, null, 2));
+  writeFile(path.join('./defaults.json'), JSON.stringify(defaults, null, 2));
 
 export default (): void => {
   const defaults = getDefaultValues();
