+++ src/interfaces/iaccent-custom-property.ts
