+++ src/interfaces/idefaults.ts
@@ -4,14 +4,10 @@ import {IAccentCustomProperty} from './iaccent-custom-property';
 export interface IDefaults {
   accents: IAccents;
   accentsProperties: IGenericObject <IAccentCustomProperty>;
-  accentableIcons: string[];
   changelog: IChangelog;
-  icons: IDefaultsThemeIcons;
   themeVariants: IDefaultsThemeVariant;
-  themeIconVariants: IDefaultsThemeIconVariant;
   themeVariantsColours: IDefaultsThemeVariant;
   themeVariantsUITheme: IDefaultsThemeVariant;
-  variantsIcons: string[];
   [Symbol.iterator](): IterableIterator<IDefaults>;
 }
 
@@ -25,31 +21,6 @@ export interface IChangelog {
   [Symbol.iterator](): IterableIterator<IChangelog>;
 }
 
-export interface IDefaultsThemeIcons {
-  theme: {
-    iconDefinitions: {
-      _folder_open: {
-        iconPath: string;
-      };
-      _folder_open_build: {
-        iconPath: string;
-      };
-      _folder_dark: {
-        iconPath: string;
-      };
-      _folder_dark_build: {
-        iconPath: string;
-      };
-      _folder_light_build: {
-        iconPath: string;
-      };
-      _folder_light: {
-        iconPath: string;
-      };
-    };
-  };
-}
-
 export interface IDefaultsThemeVariant {
   [index: string]: string;
   Darker: string;
@@ -61,12 +32,4 @@ export interface IDefaultsThemeVariant {
   PalenightHighContrast: string;
   Ocean: string;
   OceanHighContrast: string;
-}
-
-export interface IDefaultsThemeIconVariant {
-  [index: string]: string;
-  Darker: string;
-  Light: string;
-  Palenight: string;
-  Ocean: string;
 }
