+++ src/interfaces/igeneric-object.ts
