+++ src/interfaces/iinstallation-type.ts
