+++ src/interfaces/itheme-custom-properties.ts
@@ -1,4 +1,3 @@
 export interface IThemeCustomSettings {
   accent?: string;
-  autoApplyIcons?: boolean;
 }
