+++ src/webviews/ui/release-notes/style.css
